requirements.txt
alabaster==0.7.9
alembic==0.9.3
aniso8601==1.2.1
asn1crypto==0.22.0
Authlib==0.4
awscli==1.14.32
Babel==2.3.4
beautifulsoup4==4.5.3
boto3==1.5.22
botocore==1.8.36
bs4==0.0.1
cachetools==2.0.1
certifi==2018.1.18
cffi==1.9.1
chardet==3.0.4
click==6.7
colorama==0.3.7
configparser==3.5.0
cryptography==1.7.1
customized-KoNLPy==0.0.51
cycler==0.10.0
docutils==0.13.1
docx==0.2.4
EbookLib==0.16
epub==0.5.2
falcon==1.2.0
flasgger==0.6.6
Flask==0.12.2
Flask-Cache==0.13.1
Flask-Caching==1.3.3
Flask-Cors==3.0.3
Flask-Login==0.4.0
Flask-Migrate==2.1.0
Flask-MySQL==1.4.0
Flask-OAuth==0.12
Flask-OAuthlib==0.9.4
Flask-Pushjack==1.0.0
Flask-Redis==0.3.0
Flask-Session==0.3.1
Flask-SQLAlchemy==2.3.2
Flask-SSLify==0.1.5
Genshi==0.7
gevent==1.2.2
google-api-python-client==1.6.5
google-auth==1.3.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.2.0
greenlet==0.4.12
gunicorn==19.7.1
httplib2==0.10.3
idna==2.6
imagesize==0.7.1
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
JPype1==0.6.2
jsonschema==2.6.0
konlpy==0.4.4
livereload==2.5.1
lxml==4.1.1
Mako==1.0.7
Markdown==2.6.8
MarkupSafe==1.0
matplotlib==2.1.2
mistune==0.7.4
mkdocs==0.16.3
msgpack-python==0.4.8
nltk==3.2.5
ntlm-auth==1.0.5
numpy==1.13.3
oauth2==1.9.0.post1
oauth2client==4.1.2
oauthlib==2.0.6
olefile==0.44
opencv-python==3.4.0.12
ordereddict==1.1
Pillow==4.0.0
polyglot==16.7.4
protobuf==3.2.0
psycopg2==2.7.3.2
pushjack==1.3.0
py==1.4.32
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.17
Pygments==2.1.3
pylru==1.0.9
PyMySQL==0.8.0
pyOpenSSL==16.2.0
pyparsing==2.1.10
PyPDF2==1.26.0
pytest==3.0.5
pytest-runner==2.9
python-dateutil==2.6.1
python-docx==0.8.6
python-editor==1.0.3
python-markdown-math==0.3
python-mimeparse==1.6.0
python-pptx==0.6.6
pytz==2017.2
PyYAML==3.12
redis==2.10.6
requests==2.18.4
requests-ntlm==1.0.0
requests-oauthlib==0.8.0
rsa==3.4.2
s3transfer==0.1.12
scikit-learn==0.19.1
scipy==1.0.0
six==1.11.0
sklearn==0.0
snowballstemmer==1.2.1
Sphinx==1.5.1
sphinxcontrib-httpdomain==1.5.0
SQLAlchemy==1.2.3
tornado==4.4.3
uritemplate==3.0.0
urllib3==1.22
Werkzeug==0.13
XlsxWriter==1.0.2
xml-cleaner==2.0.4
xmltodict==0.11.0
yattag==1.9.2
